CARDS-2866: JWT Token Authentication Move shared code from token-authentication to new token-authentication-api module
